Reorganise controllers
