+++ backend/controllers/tasks/getTasks/index.ts
@@ -1,0 +1,10 @@
+import p from "../../../utils/prisma";
+
+export const getTasks = async (userId?: number) => {
+  const tasks = await p.task.findMany({
+    where: {
+      userId: userId,
+    },
+  });
+  return tasks;
+};
