+++ backend/controllers/users/getUser/index.ts
@@ -1,9 +1,4 @@
-import p from "../utils/prisma";
-
-export const getUsers = async () => {
-  const users = await p.user.findMany();
-  return users;
-};
+import p from "../../../utils/prisma";
 
 export const getUser = async (userId: number) => {
   const user = await p.user.findUnique({
