+++ backend/controllers/users/getUsers/index.ts
@@ -1,0 +1,7 @@
+import p from "../../../utils/prisma";
+
+export const getUsers = async () => {
+  const users = await p.user.findMany();
+  return users;
+};
+
