+++ backend/controllers/users/index.ts
@@ -1,0 +1,2 @@
+export { getUser } from "./getUser"
+export { getUsers } from "./getUsers"
