+++ backend/routers/tasks.ts
@@ -1,10 +1,10 @@
-import { router, publicProcedure } from "../utils/trpc";
 import {
-  getTaskAndChildren,
   getTaskAndAllParents,
+  getTaskAndChildren,
   getTasks,
   getTopLevelTasks,
 } from "../controllers/tasks";
+import { router, publicProcedure } from "../utils/trpc";
 import z from "zod";
 
 const tasksRouter = router({
